requirements.bot.txt
python-telegram-bot
pika